API threads & comments

+++ internal/app/thread/usecase/thread.go
@@ -1,9 +1,11 @@
 package usecase
 
 import (
+	"errors"
 	"innovaspace/internal/app/thread/repository"
 	"innovaspace/internal/domain/dto"
 	"innovaspace/internal/domain/entity"
+	"log"
 
 	"github.com/google/uuid"
 )
@@ -11,16 +13,15 @@ import (
 type ThreadUsecaseItf interface {
 	CreateThread(input dto.CreateThreadRequest) (dto.ThreadResponse, error)
 	GetAllThreads() ([]dto.ThreadResponse, error)
-	GetThreadById(threadId uuid.UUID) (dto.ThreadResponse, error)
-	UpdateThread(threadId uuid.UUID, input dto.UpdateThreadRequest) error
-	DeleteThread(threadId uuid.UUID) error
+	UpdateThread(threadId uuid.UUID, userId uuid.UUID, input dto.UpdateThreadRequest) error
+	DeleteThread(threadId uuid.UUID, userId uuid.UUID) error
 }
 
 type ThreadUsecase struct {
 	threadRepo repository.ThreadMySQLItf
 }
 
-func NewThreadUsecase(threadRepo repository.ThreadMySQL) ThreadUsecaseItf {
+func NewThreadUsecase(threadRepo repository.ThreadMySQLItf) ThreadUsecaseItf {
 	return &ThreadUsecase{
 		threadRepo: threadRepo,
 	}
@@ -65,32 +66,36 @@ func (u ThreadUsecase) GetAllThreads() ([]dto.ThreadResponse, error) {
 	return response, nil
 }
 
-func (u ThreadUsecase) GetThreadById(ThreadId uuid.UUID) (dto.ThreadResponse, error) {
-	thread, err := u.threadRepo.GetThreadById(ThreadId)
-	if err != nil {
-		return dto.ThreadResponse{}, err
-	}
-
-	return dto.ThreadResponse{
-		ThreadId: thread.ThreadId,
-		UserId:   thread.UserId,
-		Kategori: thread.Kategori,
-		Isi:      thread.Isi,
-	}, nil
-}
-
-func (u ThreadUsecase) UpdateThread(threadId uuid.UUID, input dto.UpdateThreadRequest) error {
+func (u ThreadUsecase) UpdateThread(threadId uuid.UUID, userId uuid.UUID, input dto.UpdateThreadRequest) error {
 	thread, err := u.threadRepo.GetThreadById(threadId)
 	if err != nil {
-		return err
+		return errors.New("thread not found")
+	}
+
+	if thread.UserId != userId {
+		return errors.New("not allowed to update")
 	}
 
 	thread.Kategori = input.Kategori
 	thread.Isi = input.Isi
 
+	if err := u.threadRepo.UpdateThread(thread); err != nil {
+		log.Printf("Error updating thread: %v", err)
+		return err
+	}
+
 	return u.threadRepo.UpdateThread(thread)
 }
 
-func (u ThreadUsecase) DeleteThread(threadId uuid.UUID) error {
+func (u ThreadUsecase) DeleteThread(threadId uuid.UUID, userId uuid.UUID) error {
+	thread, err := u.threadRepo.GetThreadById(threadId)
+	if err != nil {
+		return errors.New("thread not found")
+	}
+
+	if thread.UserId != userId {
+		return errors.New("not allowed to delete thread")
+	}
+
 	return u.threadRepo.DeleteThread(threadId)
 }
